back/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
Flask-Marshmallow
psycopg2
SQLAlchemy
gunicorn
python-dotenv
flask-restx